requirements.txt
mplcursors
requests
json
numpy
matplotlib.pyplot
tkinter